+++ client/src/configurator/GroupSettings.js
@@ -7,7 +7,9 @@ const _defaults = {
   namespace: [],
   id: null,
   sortOrder: 0,
-  name: ''
+  name: '',
+  alwaysShowDropdown: null,
+  defaultAlwaysShowGroupDropdowns: true
 }
 
 export default Settings.extend({
@@ -24,9 +26,8 @@ export default Settings.extend({
 
     this._templateNs = NS.parse(settings.namespace)
     this._id = settings.id
-
-    this.setSortOrder(settings.sortOrder)
-    this.setName(settings.name)
+    this._alwaysShowDropdown = settings.alwaysShowDropdown
+    this._defaultAlwaysShowGroupDropdowns = settings.defaultAlwaysShowGroupDropdowns
 
     this.$container = this._generateGroupSettings()
 
@@ -34,6 +35,10 @@ export default Settings.extend({
     this.$sortOrderInput = $neo.filter('[data-neo-gs="input.sortOrder"]')
     this.$nameInput = $neo.filter('[data-neo-gs="input.name"]')
     this.$deleteButton = $neo.filter('[data-neo-gs="button.delete"]')
+    this.$alwaysShowDropdownContainer = $neo.filter('[data-neo-gs="container.alwaysShowDropdown"]')
+
+    this.setSortOrder(settings.sortOrder)
+    this.setName(settings.name)
 
     this.addListener(this.$nameInput, 'keyup change', () => this.setName(this.$nameInput.val()))
     this.addListener(this.$deleteButton, 'click', () => {
@@ -48,7 +53,23 @@ export default Settings.extend({
     const sortOrderName = NS.fieldName('sortOrder')
     const inputId = NS.value('name', '-')
     const inputName = NS.fieldName('name')
+    const alwaysShowDropdownId = NS.value('alwaysShowDropdown', '-')
+    const alwaysShowDropdownName = NS.fieldName('alwaysShowDropdown')
     NS.leave()
+    const alwaysShowDropdownOptions = [
+      {
+        value: 'show',
+        label: Craft.t('neo', 'Show')
+      },
+      {
+        value: 'hide',
+        label: Craft.t('neo', 'Hide')
+      },
+      {
+        value: 'global',
+        label: this._defaultAlwaysShowGroupDropdowns ? Craft.t('neo', 'Use global setting (Show)') : Craft.t('neo', 'Use global setting (Hide)')
+      }
+    ]
 
     return $(`
       <div>
@@ -65,6 +86,18 @@ export default Settings.extend({
                 'data-neo-gs': 'input.name'
             }
         })}
+        <div data-neo-gs="container.alwaysShowDropdown">
+          <div class="field">
+            ${Craft.ui.createSelectField({
+              label: Craft.t('neo', 'Always Show Dropdown?'),
+              instructions: Craft.t('neo', 'Whether to show the dropdown for this group if it only has one available block type.'),
+              id: alwaysShowDropdownId,
+              name: alwaysShowDropdownName,
+              options: alwaysShowDropdownOptions,
+              value: this._alwaysShowDropdown ? 'show' : (this._alwaysShowDropdown === false ? 'hide' : 'global')
+            }).html()}
+          </div>
+        </div>
       </div>
       <hr>
       <a class="error delete" data-neo-gs="button.delete">${Craft.t('neo', 'Delete group')}</a>
@@ -92,6 +125,7 @@ export default Settings.extend({
       this._name = name
 
       this.$nameInput.val(this._name)
+      this._refreshAlwaysShowDropdown()
 
       this.trigger('change', {
         property: 'name',
@@ -99,6 +133,12 @@ export default Settings.extend({
         newValue: this._name
       })
     }
+  },
+
+  getAlwaysShowDropdown () { return this._alwaysShowDropdown },
+
+  _refreshAlwaysShowDropdown (animate) {
+    this._refreshSetting(this.$alwaysShowDropdownContainer, !!this._name, animate)
   }
 },
 {