+++ client/src/configurator/Settings.js
@@ -149,5 +149,42 @@ export default Garnish.Base.extend({
     }
 
     return attributesHtml.join('')
+  },
+
+  _refreshSetting ($container, showSetting, animate) {
+    animate = !Garnish.prefersReducedMotion() && (typeof animate === 'boolean' ? animate : true)
+
+    if (animate) {
+      if (showSetting) {
+        if ($container.hasClass('hidden')) {
+          $container
+            .removeClass('hidden')
+            .css({
+              opacity: 0,
+              marginBottom: -($container.outerHeight())
+            })
+            .velocity({
+              opacity: 1,
+              marginBottom: 24
+            }, 'fast')
+        }
+      } else if (!$container.hasClass('hidden')) {
+        $container
+          .css({
+            opacity: 1,
+            marginBottom: 24
+          })
+          .velocity({
+            opacity: 0,
+            marginBottom: -($container.outerHeight())
+          }, 'fast', () => {
+            $container.addClass('hidden')
+          })
+      }
+    } else {
+      $container
+        .toggleClass('hidden', !showSetting)
+        .css('margin-bottom', showSetting ? 24 : '')
+    }
   }
 })