+++ client/src/input/Group.js
@@ -3,6 +3,7 @@ import Garnish from 'garnish'
 const _defaults = {
   id: -1,
   sortOrder: 0,
+  alwaysShowDropdown: null,
   name: ''
 }
 
@@ -13,6 +14,7 @@ export default Garnish.Base.extend({
 
     this._id = settings.id | 0
     this._sortOrder = settings.sortOrder | 0
+    this._alwaysShowDropdown = settings.alwaysShowDropdown
     this._name = settings.name
   },
 
@@ -20,6 +22,6 @@ export default Garnish.Base.extend({
   getId () { return this._id },
   getSortOrder () { return this._sortOrder },
   getName () { return this._name },
-
+  getAlwaysShowDropdown () { return this._alwaysShowDropdown },
   isBlank () { return !this._name }
 })